impede remocao do role do usuario logado caso seja a unica a que ele possua

+++ app/Http/Controllers/Web/Admin/User/RolesController.php
@@ -133,9 +133,12 @@ class RolesController extends Controller
      * @param  Role $role
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Role $role)
     {
+        $this->authorize('admin.roles.destroy', [$role->id]);
+
         try {
 
             $role->delete();
